+++ src/util/extrapolation/StepExtrapolator.cpp
@@ -6,16 +6,16 @@
 
 #include <algorithm>
 
-void StepExtrapolator::adjust(std::vector<double> x, std::vector<double> y) {
+void StepExtrapolator::adjust(std::vector<float> x, std::vector<float> y) {
     int n = x.size();
     _slope = (y[n - 1] - y[0]) / (x[n - 1] - x[0]);
     baselineX = x[n - 1];
 }
 
-double StepExtrapolator::extrapolate(double x) {
+float StepExtrapolator::extrapolate(float x) {
     return _slope * (x - baselineX);
 }
 
-double StepExtrapolator::slope() {
+float StepExtrapolator::slope() {
     return _slope;
 }
